spug_api/requirements.txt
apscheduler==3.7.0
Django==2.2.28
asgiref==3.2.10
channels==2.3.1
channels_redis==2.4.1
paramiko==2.11.0
django-redis==4.10.0
requests==2.22.0
GitPython==3.0.8
python-ldap==3.4.0
openpyxl==3.0.3
user_agents==2.2.0
PyJWT==2.6.0
